Implemented all the things!
So it works, but with some fairly large hacks (it was made for a hack
day after all).

I think there could be some value to pulling this apart into a library
that can be used in conjuction with the Rdio SDK to make simple
visualizations, or other rhythm based things (games perhaps).